Allow builds without running the app
I would like to simply use Heroku build packs to build an app without running it if it's not needed. This change gives users the ability to use the centurylink/buildpack-runner image without "running" their app if they don't wish to.

If no extra args are passed into the run.sh script, it should inform the user and continue instead of exiting.